requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
paho-mqtt==2.1.0
aiosqlite==0.20.0
jinja2==3.1.4
weasyprint==62.3
pytest==8.3.2
httpx==0.27.0        # client async pour les tests FastAPI
python-dotenv==1.0.1
